Extract code from yearAnalysis to the separate functions
Reorder code chunks in creating plots
